+++ app/config.go
@@ -7,7 +7,7 @@ import (
 
 // ConfigHandler returns a handler function that returns API configuration
 // including Firebase public key.
-func (a *App) ConfigHandler(se GoogSystemEnv) http.HandlerFunc {
+func (a *App) ConfigHandler(se FirebaseSystemEnv) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		w.Header().Add("Content-Type", "application/json")
 		w.WriteHeader(http.StatusOK) // 200 OK