+++ schemas/postgres/schema/schema_version_function.sql
@@ -2,6 +2,6 @@ create function schema_version()
   returns text
 as
 $$
-  select 'v0.48.0' || '';
+  select 'v0.49.0' || '';
 $$
 language sql;